seed the db with event data

+++ seed.js
@@ -23,6 +23,7 @@ var chalk = require('chalk');
 var connectToDb = require('./server/db');
 var User = Promise.promisifyAll(mongoose.model('User'));
 var Venue = Promise.promisifyAll(mongoose.model('Venue'));
+var Event = Promise.promisifyAll(mongoose.model('EventProduct'));
 
 var seedUsers = function() {
 
@@ -74,6 +75,53 @@ var seedVenues = function() {
 
 };
 
+// var schema = new mongoose.Schema({
+//     name: {type: String, required: true},
+//     date: {type: Date, required: true},
+//     category: {type: String, required: true, default: 'Other'},
+//     venue: { type: mongoose.Schema.Types.ObjectId, ref: 'Venue'},
+// });
+
+
+var seedEvents = function() {
+    var venueDict={};
+    var events = [{
+        name: 'Stromae and Janelle Monae',
+        date: new Date(2015, 10, 1, 20, 0, 0),
+        venueName: 'Madison Square Garden',
+        category: 'Concert'
+    }, {
+        name: 'Hamilton',
+        date: new Date(2015,9,26,20,0,0),
+        venueName: 'Richard Rodgers Theatre',
+        category: 'Theater'
+    }, {
+        name: 'Washington Nationals at New York Mets',
+        date: new Date(2015,10,4,15,10,0),
+        venueName: 'Citi Field',
+        category: 'Sports'
+    }, {
+        name: 'Rudimental',
+        date: new Date(2015,9,29,19,0,0),
+        venueName: 'Webster Hall',
+        category: 'Concert'
+    }];
+
+    return Venue.find({}).select('name _id')
+        .then(function(venues){
+            venues.forEach(
+                function(venue){venueDict[venue.name]=venue._id;
+            });
+        })
+        .then(function(){
+            events.forEach(function(e){
+                e.venue = venueDict[e.venueName];
+                delete e.venueName;
+            });
+        })
+        .then(function(){Event.createAsync(events);});
+};
+
 connectToDb.then(function() {
     User.findAsync({})
         .then(function(users) {
@@ -93,6 +141,16 @@ connectToDb.then(function() {
                 return seedVenues();
             } else {
                 console.log(chalk.magenta('Seems to already be venue data, exiting!'));
+            }
+        })
+        .then(function() {
+            return Event.findAsync({});
+        })
+        .then(function(events){
+            if (events.length===0) {
+                return seedEvents();
+            } else {
+                console.log(chalk.magenta('Seems to already be event data, exiting!'));
                 process.kill(0);
             }
         })
